+++ src/main/java/com/dangbun/global/security/jwt/TokenAge.java
@@ -1,0 +1,4 @@
+package com.dangbun.global.security.jwt;
+
+public enum TokenAge {
+}
