+++ src/test/java/com/dangbun/domain/user/service/UserQueryServiceTest.java
@@ -10,7 +10,7 @@ import com.dangbun.domain.user.exception.custom.InvalidPasswordException;
 import com.dangbun.domain.user.exception.custom.NoSuchUserException;
 import com.dangbun.domain.user.repository.UserRepository;
 import com.dangbun.global.redis.AuthRedisService;
-import com.dangbun.global.security.refactor.JwtService;
+import com.dangbun.global.security.jwt.JwtService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -25,7 +25,6 @@ import java.util.Map;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 
 @ExtendWith(MockitoExtension.class)
